+++ src/main/java/com/woowa/DispatcherServlet.java
@@ -3,8 +3,8 @@ package com.woowa;
 import com.woowa.framework.BeanFactory;
 import com.woowa.framework.argumentresovler.ArgumentResolverComposite;
 import com.woowa.framework.web.ContentType;
-import com.woowa.framework.web.HandlerMapping;
-import com.woowa.framework.web.HandlerMethod;
+import com.woowa.framework.web.mapping.HandlerMapping;
+import com.woowa.framework.web.mapping.HandlerMethod;
 import com.woowa.framework.web.ResponseEntity;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
