+++ src/main/java/com/woowa/framework/web/HandlerMappingConfig.java
@@ -2,6 +2,8 @@ package com.woowa.framework.web;
 
 import com.woowa.framework.Bean;
 import com.woowa.framework.BeanFactory;
+import com.woowa.framework.web.mapping.DynamicHandlerMapping;
+import com.woowa.framework.web.mapping.StaticResourceHandlerMapping;
 
 public class HandlerMappingConfig {
