+++ src/main/java/com/woowa/framework/web/mapping/HandlerMethod.java
@@ -1,5 +1,6 @@
-package com.woowa.framework.web;
+package com.woowa.framework.web.mapping;
 
+import com.woowa.framework.web.ResponseEntity;
 import java.lang.reflect.Parameter;
 
 public interface HandlerMethod {
